CNPM/Hotel_Booking_App/requirements.txt
flask
mysql.connector
flask_jwt_extended
flask_cors
dotenv
ssl